+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2026-08-09
+
+### Added
+
+- Added secret switch to enable/disable "Slim Mode"
+- Added skills
+- Added reference to full CV when printing to PDF
+- Added secret function to swap location easily
+- Added secret option to remove sub-sections that are not normally removable
+- Minor Fixes and Improvements
+
 ## [1.3.0] - 2026-07-30
 
 ### Added
@@ -43,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [START of the PROJECT]
 - Initial project setup and configuration.
 
-[unreleased]: https://github.com/filippochinni/Curriculum-Vitae/compare/v1.3.0...HEAD
+[unreleased]: https://github.com/filippochinni/Curriculum-Vitae/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/filippochinni/Curriculum-Vitae/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/filippochinni/Curriculum-Vitae/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/filippochinni/Curriculum-Vitae/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/filippochinni/Curriculum-Vitae/compare/v1.0.0...v1.1.0